requirements.txt
playwright
pytest
allure-pytest
pytest-playwright